- **Fix:** Standardized `cancel` API to accept `Object` type for notification IDs across platforms. - **Feat:** Introduced iOS background task management for handling pending notifications. - **Update:** Migrated platform dependencies to Git-based sources for better version control. - **Chore:** Added SonarQube workflow and `CODEOWNERS` file for improved code quality and review process. - **Fix:** Resolved crash when handling notification icons and payload extraction on Android.

+++ flutter_local_notifications/example/lib/main.dart
@@ -3231,7 +3231,7 @@ class _HomePageState extends State<HomePage> {
                       child: const Text('Get messaging style'),
                       onPressed: () {
                         _getActiveNotificationMessagingStyle(
-                          activeNotification.id!,
+                          activeNotification.id! as int,
                           activeNotification.tag,
                         );
                       },
